backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
attrs==25.4.0
bcrypt==5.0.0
blinker==1.9.0
boto3==1.40.64
botocore==1.40.64
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.1
dill==0.4.0
dnspython==2.8.0
docker==7.1.0
fastapi==0.120.4
Flask==3.0.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
google-generativeai==0.8.3
google-pasta==0.2.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
gunicorn==22.0.0
h11==0.16.0
idna==3.11
importlib-metadata==6.11.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
maptiler==1.0.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mock==4.0.3
multiprocess==0.70.18
numpy==2.2.6
omegaconf==2.3.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==24.2
pandas==2.3.3
passlib==1.7.4
pathos==0.3.4
pillow==12.0.0
platformdirs==4.5.0
pox==0.3.6
ppft==1.7.7
protobuf==6.31.1
psutil==7.1.2
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.10.1
python-dateutil==2.9.0.post0
weaviate-client==4.10.0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.28.0
s3transfer==0.14.0
sagemaker==2.254.1
sagemaker-core==1.0.61
schema==0.7.8
six==1.17.0
smdebug-rulesconfig==1.0.1
sniffio==1.3.1
starlette==0.49.1
tblib==3.2.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
Werkzeug==3.1.3
zipp==3.23.0
